+++ app/src/main/java/com/example/gajamap/ui/fragment/setting/KakaoProfileFragment.kt
@@ -29,7 +29,7 @@ class KakaoProfileFragment: BaseFragment<FragmentKakaoProfileBinding>(R.layout.f
 
     // 선택된 클라이언트들을 저장하기 위한 리스트
     private var selectedClients: MutableList<Clients?> = mutableListOf()
-    private var groupId : Int = -1
+    private var groupId : Long = -1
     override val viewModel by viewModels<ClientViewModel> {
         ClientViewModel.SettingViewModelFactory("tmp")
     }
